IGNITE-24384 / move a class
